# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: HypeCheck API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List or search products
      parameters:
        - in: query
          name: q
          required: false
          schema:
            type: string
          description: Search query (name, category, or retailer)
        - in: query
          name: limit
          required: false
          schema:
            type: number
            default: 20
            minimum: 1
            maximum: 100
        - in: query
          name: offset
          required: false
          schema:
            type: number
            default: 0
            minimum: 0
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProductSummary"
                  total:
                    type: number
        "502":
          description: Supabase unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product detail with aggregate fields
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "502":
          description: Supabase unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /products/{productId}/reviews:
    get:
      operationId: getProductReviews
      tags: [products]
      summary: Get reviews for a product
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: number
            default: 20
            minimum: 1
            maximum: 100
        - in: query
          name: offset
          required: false
          schema:
            type: number
            default: 0
            minimum: 0
      responses:
        "200":
          description: Product reviews
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Review"
                  total:
                    type: number
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "502":
          description: Supabase unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"

  # ---------------------------------------------------------------------------
  # Reviews
  # ---------------------------------------------------------------------------
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get the current user's profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "502":
          description: Supabase unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /profile/reviews:
    get:
      operationId: getProfileReviews
      tags: [profile]
      summary: Get reviews authored by the current user
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: number
            default: 20
            minimum: 1
            maximum: 100
        - in: query
          name: offset
          required: false
          schema:
            type: number
            default: 0
            minimum: 0
      responses:
        "200":
          description: User reviews
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Review"
                  total:
                    type: number
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "502":
          description: Supabase unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"

  # ---------------------------------------------------------------------------
  # Media
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorBody:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string

    # -------------------------------------------------------------------------
    # Auth
    # -------------------------------------------------------------------------
...
    ProductSummary:
      type: object
      required: [id, name, category, price, reviewCount]
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        price:
          type: number
          description: Retail price
        retailer:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
          description: Average rating (1–5), null when no reviews yet
        reviewCount:
          type: number
        worthTheHype:
          type: number
          nullable: true
          description: Percentage (0–100) of reviewers who said it is worth the hype
    ProductDetail:
      type: object
      required: [id, name, category, price, reviewCount]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          description: Retail price
        category:
          type: string
        retailer:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        storeUrl:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
          description: Average rating (1–5), null when no reviews yet
        reviewCount:
          type: number
        worthTheHype:
          type: number
          nullable: true
          description: Percentage (0–100) of reviewers who said it is worth the hype

    # -------------------------------------------------------------------------
    # Reviews
    # -------------------------------------------------------------------------
...
    Review:
      type: object
      required: [id, productId, userId, rating, worthIt, reviewText, createdAt]
      properties:
        id:
          type: string
        productId:
          type: string
        userId:
          type: string
        rating:
          type: number
          minimum: 1
          maximum: 5
        worthIt:
          type: boolean
        reviewText:
          type: string
        videoUrl:
          type: string
          nullable: true
        photoUrl:
          type: string
          nullable: true
        createdAt:
          type: string
        authorUsername:
          type: string
          nullable: true
        authorAvatarUrl:
          type: string
          nullable: true

    # -------------------------------------------------------------------------
    # Profile
    # -------------------------------------------------------------------------
    UserProfile:
      type: object
      required: [id, email, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        username:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string

    # -------------------------------------------------------------------------
    # Media
    # -------------------------------------------------------------------------
...