# lib/api-zod/src/generated/types/productSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * HypeCheck API specification
 * OpenAPI spec version: 0.2.0
 */

export interface ProductSummary {
  id: string;
  name: string;
  category: string;
  /** Retail price */
  price: number;
  retailer?: string | null;
  imageUrl?: string | null;
  /** Average rating (1–5), null when no reviews yet */
  rating?: number | null;
  reviewCount: number;
  /** Percentage (0–100) of reviewers who said it is worth the hype */
  worthTheHype?: number | null;
}