# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * HypeCheck API specification
 * OpenAPI spec version: 0.2.0
 */

export interface Review {
  id: string;
  productId: string;
  userId: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  worthIt: boolean;
  reviewText: string;
  videoUrl?: string | null;
  photoUrl?: string | null;
  createdAt: string;
  authorUsername?: string | null;
  authorAvatarUrl?: string | null;
}