# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * HypeCheck API specification
 * OpenAPI spec version: 0.2.0
 */

export interface UserProfile {
  id: string;
  email: string;
  username?: string | null;
  avatarUrl?: string | null;
  createdAt: string;
}